tetragon: Add generic_retkprobe function
